A LOT of stuff was changed after realizing that I need more room for other actions/reducers and a create challenge page for new users. Reiterating that the actual status bar has not been implemented yet and instead changed name to LeftContainer since this will hold multiple components including challenger container for the status bar as well as books finished
