+++ client/src/pages/SignUpPage.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { signup} from "../store/auth";
+import { signup} from "../store/actions/auth";
 import { Redirect, Link } from "react-router-dom";
 import {Container} from "@material-ui/core";
 import {makeStyles } from "@material-ui/core/styles"
