+++ client/src/store/reducers/rootReducer.js
@@ -1,0 +1,9 @@
+import {combineReducers } from 'redux';
+import authReducer from './authReducer';
+//import challengeReducer from './challengeReducer';
+
+const rootReducer = combineReducers({
+    auth: authReducer
+})
+
+export default rootReducer;
